\name{perf_collections}
\alias{perf_collections}
\title{Get quarterly CX data collections summary}
\description{
Returns a tibble of quarterly HISP data collection summaries
with universe size, survey volume, and response rates.
}
\usage{
perf_collections()
}
\value{
A tibble of quarterly collection summaries.
}
