\name{perf_context}
\alias{perf_context}
\title{Get performance.gov client source and function signatures}
\description{
Reads and returns the source code of all public functions
defined in this client file.
}
\usage{
perf_context()
}
\value{
Character string of function source (invisibly), also printed.
}
